refactor(tests): reorganize test files and update import paths
Move test files from src/ to dedicated __tests__ directories for runtime and type tests,
updating import statements to use relative paths from new locations
